Refactor : 중간 커밋
