+++ sport_companion/module-api/src/main/java/com/service/sport_companion/api/component/vote/CandidateHandler.java
@@ -1,4 +1,4 @@
-package com.service.sport_companion.api.component;
+package com.service.sport_companion.api.component.vote;
 
 import com.service.sport_companion.domain.entity.CandidateEntity;
 import com.service.sport_companion.domain.model.dto.response.vote.CandidateAndCountDto;
