+++ sport_companion/module-api/src/main/java/com/service/sport_companion/api/service/impl/VoteServiceImpl.java
@@ -1,8 +1,8 @@
 package com.service.sport_companion.api.service.impl;
 
-import com.service.sport_companion.api.component.CandidateHandler;
-import com.service.sport_companion.api.component.UserVoteHandler;
-import com.service.sport_companion.api.component.VoteHandler;
+import com.service.sport_companion.api.component.vote.CandidateHandler;
+import com.service.sport_companion.api.component.vote.UserVoteHandler;
+import com.service.sport_companion.api.component.vote.VoteHandler;
 import com.service.sport_companion.api.service.VoteService;
 import com.service.sport_companion.core.exception.GlobalException;
 import com.service.sport_companion.domain.entity.CandidateEntity;
@@ -35,6 +35,9 @@ public class VoteServiceImpl implements VoteService {
   private final CandidateHandler candidateHandler;
   private final UserVoteHandler userVoteHandler;
 
+  /**
+   * 투표 생성 (관리자 기능)
+   */
   @Override
   public ResultResponse<Void> createVote(Long userId, CreateVoteDto voteDto) {
     // 동일한 날짜에 이미 등록돼있으면 추가로 등록할 수 없음
@@ -65,6 +68,9 @@ public class VoteServiceImpl implements VoteService {
     return ResultResponse.of(SuccessResultType.SUCCESS_CREATE_VOTE);
   }
 
+  /**
+   * 투표 수정 (관리자 기능)
+   */
   @Override
   public ResultResponse<Void> updateVote(Long userId, Long voteId, CreateVoteDto voteDto) {
     // 투표 주제 업데이트
@@ -89,6 +95,9 @@ public class VoteServiceImpl implements VoteService {
     return ResultResponse.of(SuccessResultType.SUCCESS_MODIFY_VOTE);
   }
 
+  /**
+   * 투표 삭제 (관리자 기능)
+   */
   @Override
   public ResultResponse<Void> deleteVote(Long userId, Long voteId) {
     candidateHandler.deleteVoteByVoteId(voteId);
@@ -97,6 +106,9 @@ public class VoteServiceImpl implements VoteService {
     return ResultResponse.of(SuccessResultType.SUCCESS_DELETE_VOTE);
   }
 
+  /**
+   * 이번주 투표 조회
+   */
   @Override
   public ResultResponse<VoteResponse> getThisWeekVote() {
     // 이번주 투표 시작일인 월요일의 날짜 찾기
@@ -110,6 +122,9 @@ public class VoteServiceImpl implements VoteService {
     );
   }
 
+  /**
+   * 이번주 투표 결과 조회
+   */
   @Override
   public ResultResponse<VoteResponse> getThisWeekVoteResult(Long userId) {
     // 이번주 투표와 결과 조회
@@ -125,6 +140,9 @@ public class VoteServiceImpl implements VoteService {
     );
   }
 
+  /**
+   * 이전 투표 및 결과 조회 - 최신순, 인기순
+   */
   @Override
   public ResultResponse<PageResponse<VoteResponse>> getPrevVoteResult(
     Long userId, SortType sortType, Pageable pageable
@@ -152,6 +170,9 @@ public class VoteServiceImpl implements VoteService {
       ));
   }
 
+  /**
+   * 사용자 당일 투표 여부 조회
+   */
   @Override
   public ResultResponse<CheckVotedResponse> checkUserVoted(Long userId) {
     return new ResultResponse<>(
@@ -160,6 +181,9 @@ public class VoteServiceImpl implements VoteService {
     );
   }
 
+  /**
+   * 사용자가 이번주 투표의 후보에 투표
+   */
   @Override
   public ResultResponse<VoteResponse> vote(Long userId, Long voteId, Long candidateId) {
     // 오늘 투표한 적 있는지 확인
@@ -188,13 +212,17 @@ public class VoteServiceImpl implements VoteService {
     );
   }
 
-  // 투표 시작 날짜 구하기
+  /**
+   * 투표 시작 날짜 구하기
+   */
   private LocalDate getVoteStartDate(DayOfWeek dayOfWeek) {
     return LocalDate.now()
       .with(TemporalAdjusters.previousOrSame(dayOfWeek));
   }
 
-  // 사용자가 후보 List 중에서 투표한 곳을 조회
+  /**
+   * 사용자가 후보 List 중에서 투표한 곳을 조회
+   */
   private Long getUserVotedCandidateId(Long userId, List<CandidateAndCountDto> candidateList) {
     if (userId == null) {
       return null;
@@ -209,7 +237,9 @@ public class VoteServiceImpl implements VoteService {
     );
   }
 
-  // 지난 투표를 원하는 정렬 조건에 맞는 순서로 pageable 크기만큼 조회
+  /**
+   * 지난 투표를 원하는 정렬 조건에 맞는 순서로 pageable 크기만큼 조회
+   */
   private Page<VoteEntity> getPrevVoteListBySortType(SortType sortType, Pageable pageable) {
     if (sortType.equals(SortType.PARTICIPANT)) {
       return voteHandler.findPrevVoteOrderByParticipant(getVoteStartDate(DayOfWeek.MONDAY), pageable);
