+++ sport_companion/module-domain/src/main/java/com/service/sport_companion/domain/model/dto/response/PageResponse.java
@@ -12,4 +12,5 @@ public class PageResponse<T> {
   private int totalPage;
   private Long totalData;
   private List<T> dataList;
+
 }
